MULE-16140: Source On Error callback is not called when an error occurs (#7526)
(cherry picked from commit 044aa4e0d7f2b586dabb32f93b7bd5aa6471e195)

This is only a partial cherry pick of the commit in support/4.1.5. This
was intended only to build an aggregate patch. The rest of the changes
are obtained from the original patch when packaging the new aggregate
one.